refactor FileSystem monitors
